ZentangleTask that evolved random PicBreeder Images
@sarahkfriday @krolikoa If you run the new ZentangleTask, it will keep spitting out random PicBreeder images. It makes a random population, assigns each one a random score, creates children from the parents, and then picks the best images from the combined child/parent population (however, best is defined by having the highest random number).

This is a good base for setting up automated experiments. The next step for me is to use my knowledge of how the different evolution tasks are defined to allow a zentangle (perhaps several?) each generation.

Something I might as one of you to try later is to define actual fitness functions, instead of this pure random selection, but that would just be a bonus assuming we even have the time.

Note that nearly all of the other files included in this commit are simply things that are being copied back from MMNEAT with little change ... a few unsupported features are commented out here and there.

One step toward #6