Another fitness function
This one has fitness functions for half-black, but also encourages each of the 3 color channels individually. Unfortunately, this means that some images crop up that are almost pure blue, red, or green. It's ok when they mix, but there can still be boring results.

I guess the fact that automating this process is difficult could be viewed as a good thing, but it seems that the random fitness function may actually produce the best results over time.