Single pattern degenerates no longer possible
I think this makes single pattern degenerates impossible. I actually found a small bug when investigating what turned out to be some incorrect technical details in the paper. Black vs White is determined by first calculating the max/min brightness values. The idea is clearly that values closer to the max are white and others are black, and this should obviously depend on a calculation of the midpoint ... but the code was using minus instead of plus.

Although this might make single pattern degenerates impossible, There are still some degenerates that are a boring half/half split. Re-running automated experiments now.

@krolikoa @sarahkfriday